requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
boto3=1.34.79=pypi_0
botocore=1.34.79=pypi_0
ca-certificates=2024.8.30=hbcca054_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
cheroot=10.0.1=pypi_0
clip=0.2.0=pypi_0
cn-clip=1.5.1=dev_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
faiss-gpu=1.7.2=pypi_0
filelock=3.13.3=pypi_0
fonttools=4.51.0=pypi_0
fsspec=2024.3.1=pypi_0
ftfy=6.2.0=pypi_0
google-auth=2.32.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.65.0=pypi_0
h5py=3.8.0=pypi_0
huggingface-hub=0.23.3=pypi_0
idna=3.6=pypi_0
importlib-metadata=8.0.0=pypi_0
importlib-resources=6.4.0=pypi_0
jaraco-functools=4.0.1=pypi_0
jinja2=3.1.3=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.5=pypi_0
langid=1.1.6=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc=14.1.0=h77fa898_1
libgcc-ng=14.1.0=h69a702a_1
libgomp=14.1.0=h77fa898_1
libstdcxx=14.1.0=hc0a3c3a_1
libstdcxx-ng=14.1.0=h4852527_1
lmdb=1.3.0=pypi_0
markdown=3.6=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
more-itertools=10.3.0=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numpy=1.23.0=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.3.2=hb9d3cd8_0
packaging=24.1=pypi_0
pandas=2.0.3=pypi_0
pillow=10.3.0=pypi_0
pip=23.3.1=py38h06a4308_0
protobuf=5.27.2=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pyparsing=3.1.2=pypi_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
s3transfer=0.10.1=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.2.2=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=3.5.0=pypi_0
timm=1.0.7=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.19.1=pypi_0
torch=1.8.1+cu111=pypi_0
torchvision=0.9.1+cu111=pypi_0
tqdm=4.66.2=pypi_0
transformers=4.41.2=pypi_0
triton=2.3.1=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=1.26.18=pypi_0
wcwidth=0.2.13=pypi_0
web-py=0.62=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.41.2=py38h06a4308_0
whl=0.0.4=pypi_0
xz=5.4.6=h5eee18b_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_0
